Improve cycle detection
Keep a set of any bundle we've inspected before as a different set
compared to the current bundles being inspected as part of the depth
first chain.

This fixes a major performance problem with re-evaluating bundles that
have already been seen before while still allowing cycles to be
spotted and accurately reported in the bundle configurations.

Signed-off-by: William Douglas <william.r.douglas@gmail.com>